requirements.txt
pandas
laspy
mysql-connector-python